gpu_faster_than_cpu

+++ matrix_mult.cl
@@ -1,7 +1,7 @@
 __kernel void multiply(
-    __global const double* A,    // Macierz wejściowa A (n x k)
-    __global const double* B,    // Macierz wejściowa B (k x m)
-    __global double* C,          // Macierz wyjściowa C (n x m)
+    __global const float* A,    // Macierz wejściowa A (n x k)
+    __global const float* B,    // Macierz wejściowa B (k x m)
+    __global float* C,          // Macierz wyjściowa C (n x m)
     const int n,                 // Liczba wierszy A
     const int k,                 // Liczba kolumn A / wierszy B
     const int m)                 // Liczba kolumn B
@@ -13,7 +13,7 @@ __kernel void multiply(
 
     // Upewniamy się, że nie wychodzimy poza granice macierzy
     if (i < n && j < m) {
-        double sum = 0.0;
+        float sum = 0.0;
         for (int l = 0; l < k; l++) {
             // C[i][j] += A[i][l] * B[l][j]
             sum += A[i * k + l] * B[l * m + j];
